Ensure we only find in the registry each bundle once
- Fix race conditions
- Refactor some tests
